sabnzbd: use scoop version of UnRAR.exe & MultiPAR
Only if SABnzbd is being installed globally and on 64-bit.
This might change in the future, but right now I have newer versions
installed than what SAB are distributing.

If running into problems with unrarring or repairing, please restore
the original EXEs first before making a bug report.

7za is not substituted because there apparently is no 7za being
distributed in the offical Scoop buckets and I don't want to replace
it with 7z.exe, which can load external DLLs to provide archive support.